+++ packages/react-container/src/components/index.ts
@@ -5,3 +5,4 @@ export * from './Sentry';
 export * from './TranslationWrapper';
 export * from './Intercom';
 export * from './Logout';
+export * from './LogoutButton';